+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+# [0.4.4] - 2023-08-01
+
+###
+
+- Added Makefile and install.sh in src folder to allow 
+  build via make.
+
+- Fixed issue where start_ebpf_router.py was not   
+  properly updating the ziti-router.service file.
+
 # [0.4.3] - 2023-07-25
 
 ###